User-Post connection

+++ controllers/feed.js
@@ -4,6 +4,7 @@ const path = require("path")
 const { validationResult } = require("express-validator")
 
 const Post = require("../models/post")
+const User = require("../models/user")
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1
@@ -14,6 +15,7 @@ exports.getPosts = (req, res, next) => {
         .then((count) => {
             totalItems = count
             return Post.find()
+                .populate({ path: 'creator', select: 'name'})
                 .skip((currentPage - 1) * perPage)
                 .limit(perPage)
         })
@@ -47,6 +49,7 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path.replace("\\", "/")
     const title = req.body.title
     const content = req.body.content
+    let creator
     const post = new Post({
         title: title,
         content: content,
@@ -55,10 +58,19 @@ exports.createPost = (req, res, next) => {
     })
     post.save()
         .then((result) => {
-            
+            return User.findById(req.userId)
+        })
+        .then((user) => {
+            creator = user
+            user.posts.push(post)
+            return user.save()
+        })
+        .then((result) => {
+            console.log(creator)
             res.status(201).json({
                 message: "Post created successfully!",
-                post: result,
+                post: post,
+                creator: { _id: creator._id, name: creator.name },
             })
         })
         .catch((err) => {
@@ -114,6 +126,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404
                 throw error // move to catch error
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error("Not authorized!")
+                error.statusCode = 403
+                throw error
+            }
             if (imageUrl != post.imageUrl) {
                 clearImage(post.imageUrl)
             }
@@ -145,12 +162,23 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404
                 throw error // move to catch error
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error("Not authorized!")
+                error.statusCode = 403
+                throw error
+            }
             // Check logged in user
             clearImage(post.imageUrl)
             return Post.findByIdAndRemove(postId)
         })
         .then((result) => {
-            console.log(result)
+            return User.findById(req.userId)
+        })
+        .then(user => {
+            user.posts.pull(postId)
+            return user.save()
+        })
+        .then(result => {
             res.status(200).json({ message: "Deleted post." })
         })
         .catch((err) => {
